# Optimize Newton-Schulz Iteration in `zeropower_via_newtonschulz5` Function