--- a/muon.py
+++ b/muon.py
@@ -14,15 +14,17 @@ def zeropower_via_newtonschulz5(G, steps=10, eps=1e-7):
     performance at all relative to UV^T, where USV^T = G is the SVD.
     """
     assert len(G.shape) == 2
-    a, b, c = (3.4445, -4.7750,  2.0315)
-    X = G.bfloat16()
-    X /= (X.norm() + eps) # ensure top singular value <= 1
+    a, b, c = (3.4445, -4.7750, 2.0315)
+    # X = G.bfloat16()
+    X = G.float()
+    X /= (X.norm() + eps)
     if G.size(0) > G.size(1):
         X = X.T
     for _ in range(steps):
         A = X @ X.T
-        B = b * A + c * A @ A
-        X = a * X + B @ X
+        Y = A @ X
+        Z = A @ Y
+        X = a * X + b * Y + c * Z
     if G.size(0) > G.size(1):
         X = X.T
     return X