--- a/muon.py
+++ b/muon.py
@@ -15,8 +15,7 @@ def zeropower_via_newtonschulz5(G, steps=10, eps=1e-7):
     """
     assert len(G.shape) == 2
     a, b, c = (3.4445, -4.7750, 2.0315)
-    # X = G.bfloat16()
-    X = G.float()
+    X = G.bfloat16()
     X /= (X.norm() + eps)
     if G.size(0) > G.size(1):
         X = X.T