@KellerJordan 2024-12-04
I'm confused how this saves a matmul, looks like there are still 3?