@KellerJordan 2024-12-04
the timings r also confusing; why would shape [128, 256] have worse time_original than shape [16384, 32K] ?