@Dune-Z 2024-12-04
> I'm confused how this saves a matmul, looks like there are still 3?

Thanks for your reply. If I understand it correctly, the original implementation computes:

```
A2 = A @ A             # 1 matmul
Z1 = A @ X             # 1 matmul
Z2 = A2 @ X            # 1 matmul
X_new = a * X + b * Z1 + c * Z2
```

The modifed implementation computes:
```
Z1 = A @ X             # 1 matmul
Z2 = A @ Z1            # 1 matmul
X_new = a * X + b * Z1 + c * Z2
```

We save one matmul by reusing the result `Z1`. Assume the dimension of X is $n\times m, m > n$. Then the original one has $n^3 + 2n^2m$ operations and the modified one has $2n^2m$.