@Dune-Z 2024-12-05
> if I'm reading this diff correctly, two matmuls were removed and two were added https://github.com/KellerJordan/Muon/pull/6/files#diff-ff0575a769b2390ce3256edb1c20e4d741d514a77c4f0697c2fa628f810f46b1L24

`B @ X` contains `A@X` and `(A@A)@X` two matmul.