@Dune-Z 2024-12-05
> if I'm reading this diff correctly, two matmuls were removed and two were added https://github.com/KellerJordan/Muon/pull/6/files#diff-ff0575a769b2390ce3256edb1c20e4d741d514a77c4f0697c2fa628f810f46b1L24

You are right about the number of matmul, my previous explanation is incorrect. Thank you for pointing out the profiling results. I will take some time to investigate further to determine whether they indeed provide a benefit or if there might be an issue with the profiling itself. I'll close the request to avoid potentially misleading others.